+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [[v0.0.4]](https://github.com/mlange-42/arche-model/compare/v0.0.4...v0.0.5)
+## [[v0.0.5]](https://github.com/mlange-42/arche-model/compare/v0.0.4...v0.0.5)
+
+### Breaking changes
+
+* Renamed `Systems.Fps` and `Systems.Tps` to `Systems.FPS` and `Systems.TPS` (#26)
 
 ### Features